chore(git): tell git to use lf even in windows
The repo is configured to use LF for line endings. But when
checking it out on windows (e.g. to run E2E tests on CI) git will,
by default, use CRLF. This breaks the CI on windows due to
triggering prettier during linting.

Normally prettier or the related config defined in .editorconfig
would be enough to cover this. But if a windows-using contributor
commits a new file, it would be allowed to sneak through and be
comitted with CRLF while all other files use LF.

This commit works around prettier seeing different line-endings
than expected, fixes the CI on windows env  and patches that issue.

Signed-off-by: Konstantinos Maninakis <maninak@protonmail.com>