ci(repo): set-up github actions for pull requests (#36)
Also fixes an end-of-line marker issue:

The repo is configured to use LF for line endings. But when
checking it out on Windows (e.g. to run E2E tests on CI) git will,
by default, use CRLF. This breaks the CI on windows due to
triggering prettier during the lining check.

Normally prettier or the related config defined in .editorconfig
would be enough to cover this. But if a Windows-using contributor
commits a new file, it would be allowed to sneak through and be
committed with CRLF while all other files use LF.

This commit works around prettier seeing different line-endings
than expected, fixes the CI on Windows env and patches that
issue.

Co-authored-by: Yorgos Saslis <2420882+gsaslis@users.noreply.github.com>
Signed-off-by: Kostis Maninakis <maninak@users.noreply.github.com>