feat(app): [wip] retrieve all patches with `rad cob list|show`
Current implementation is much slower than with httpd!

Transformation function is not complete:
- not all timestamps have been normalized
- activity section in Patch Detail view seems to have an bug
- "augmented patch" operations should be all done in
the same transformation func

In patchStore the refreshPatch should be updated to also not use httpd.

Aliases need to be seeded from `rad follow`'s output.

It feels like if I could query the db directly instead, it may all be
much better, but alas, some module resolution issue with the
two main sqlite libs stands in the way...

Signed-off-by: Konstantinos Maninakis <maninak@protonmail.com>